+++ src/ListenerRemovableBasedOnHandleCount.php
@@ -1,0 +1,12 @@
+<?php declare(strict_types=1);
+
+namespace Labrador\AsyncEvent;
+
+/**
+ * @api
+ */
+interface ListenerRemovableBasedOnHandleCount {
+
+    public function handleLimit() : int;
+
+}